Use small file feature of global sort to minimise the carbondata file count
